labs/iot/gps-lab/code/server/requirements.txt
flask
flask_socketio
azure.eventhub
python-dotenv